td-api/requirements.txt
tda-api
selenium

psycopg2

